refactor: coupon api 컨트롤러 통합
